Moved terminology config out of being R4 specific. It now applies to all versions
